// src/core/redux/type/nft.type.ts
export interface CollectionMoralisResponse {
  data: CollectionMoralisType;
  isSuccess: boolean;
  status: number;
}
// ...
interface CollectionMoralisType {
  cursor: string;
  page: number;
  page_size: number;
  result: UnAddedNFTListType[];
  status: string;
}

export type NFTTonItems = NftItem & {
  active?: boolean;
};

export type TonNFTResponses = {
  nftItems: NFTTonItems[];
};

export type NFTSlice = {
  tonCollection: {
    [idAccount: string]: {
      [idAddress: string]: NFTTonCollection[];
    };
  };
  collection: {
    [idAccount: string]: {
      [idAddress: string]: NFTCollection[];
    };
  };
  collectionsMoralis: CollectionMoralisResponse | {};
  detailNFTsByCollection: NFTDetailColectionAPIResponse | {};
  cursorCollectionsMoralis: string | null;
  cursorDetailNFTsByCollection: string | null;
  currentRoutesTabIndex: number;
  archivedCollectionFocused: string[];
};

export type ImportNFTParams = {
  contractAddress: string;
  nftId: number;
  id: string;
  protocolData: ProtocolDataFromBEType;
  accountId: string;
  nftData: NFTResponse;
};
export type NFTResponse = {
  name: string;
  owner: string;
  symbol: string;
  tokenURI: string;
  tokenStandard: NFTTokenStandard;
  quantity?: number;
};

export type NFTTonResponse = {
  name: string;
  owner: string;
  symbol: string;
  tokenURI?: string;
  tokenStandard?: NFTTokenStandard;
  quantity?: number;
};

export type ImportNFTTonParams = {
  contractAddress: string;
  id: string;
  protocolData: ProtocolDataFromBEType;
  accountId: string;
  dataTonNFT: NftTonNewItem;
};
export type NFTRootData = Omit<NFTResponse, "quantity" | "tokenStandard"> & {
  contractAddress: string;
  id: string;
  accountId: string;
  protocol: ProtocolDataFromBEType;
};
export type NFTCollection = {
  name: string;
  contractAddress: string;
  network: string;
  tokenStandard?: NFTTokenStandard;
  data: NFTData[];
  latestId?: number; // only for ERC1155
};
export type NFTData = {
  root: NFTRootData;
  detail: NFTType;
};
// ...
export type MigrateNFTType = {
  accountId: string;
  chainId: number;
  walletAddress: string;
  idCollection: string;
};

export type NFTDetailColectionAPIResponse = {
  cursor: string;
  page: number;
  page_size: number;
  result: NFTDetailEVMCollectionType[];
};

export interface CollectionDetailMoralisResponse {
  data: NFTDetailColectionAPIResponse;
  isSuccess: boolean;
  status: number;
  result: NFTDetailEVMCollectionType[];
  cursor: string | null;
}

export interface NFTDetailEVMCollectionType {
  token_address: string;
  token_id: string;
  amount: string;
  token_hash: string;
  block_number: string;
  block_number_minted: null | string;
  contract_type: string;
  name: string;
  symbol: string;
  token_uri: string;
  metadata: string;
  last_token_uri_sync: string;
  last_metadata_sync: string;
  minter_address: null | string;
  owner_of: string;
  rarity_rank: number;
  rarity_percentage: number;
  rarity_label: string;
  possible_spam: boolean;
  verified_collection: boolean;
  floor_price: null | string;
  floor_price_usd: null | string;
  floor_price_currency: null | string;
  active: boolean;
}

//////////////////////////////////////////
// TON

export type NFTDetailTonColectionType = NFTDetailEVMCollectionType &
  NftItem & {
    active?: boolean;
  };

export type NFTTonRootData = Omit<NFTTonResponse, "quantity"> & {
  contractAddress: string;
  id: string;
  accountId: string;
  protocol: ProtocolDataFromBEType;
  tonCollectionAddress: string;
};

export type NFTTonType = {
  description?: string;
  external_url?: string;
  image?: string;
  name?: string;
  attributes?: [];
  network_image?: string;
  image_data?: string;
  quantity?: number;
  tokenStandard?: NFTTokenStandard;
  nftDetailAll: NftTonNewItem;
};

export type NFTTonData = {
  root: NFTTonRootData;
  detail: NFTTonType;
};

export type NFTTonCollection = {
  name: string;
  contractAddress: string;
  network: string;
  data: NFTTonData[];
  latestId?: number;
};

export type NFTType = {
  description?: string;
  external_url?: string;
  image: string;
  name: string;
  attributes?: [];
  network_image?: string;
  nftId: number;
  image_data?: string;
  quantity?: number;
  tokenStandard?: NFTTokenStandard;
  nftDetailAll: NFTDetailEVMCollectionType & NftItem;
};